requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
pillow==10.4.0
torch
torchvision
